SpreadWidgets: don't be evil as gtk, scale is a double!
(bzr r4163.7.8)

+++ unity-shared/SpreadWidgets.cpp
@@ -154,7 +154,7 @@ public:
     }
   }
 
-  int scale_;
+  double scale_;
   dash::BaseTexturePtr corner_tex_;
   dash::BaseTexturePtr left_edge_tex_;
   dash::BaseTexturePtr horizontal_tex_;
